V V

+++ libft/ft_strlcat.c
@@ -6,7 +6,7 @@
 /*   By: csakamot <csakamot@student.42tokyo.jp>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/05/14 01:50:50 by yayoi             #+#    #+#             */
-/*   Updated: 2023/05/16 16:14:41 by csakamot         ###   ########.fr       */
+/*   Updated: 2023/05/17 16:08:21 by csakamot         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,7 +16,7 @@ size_t	ft_strlcat(char *dest, const char *src, size_t size);
 
 size_t	ft_strlcat(char *dest, const char *src, size_t size)
 {
-	int		i;
+	size_t		i;
 	size_t	d_len;
 	size_t	s_len;
